Fix conda environment so application runs
This commit fixes the conda environment simply not building, caused by the results of the conda solver changing over time as new packages were released.

I've manually gone in and made every major dependency explicit with comments documenting my changes to not only fix the issue but to try and prevent this from happening again.

I chose to not make it the environment.yaml an env export to make it easy to update in the future if changes are made to this codebase.